requirements.txt
kubernetes
PyYAML
click